Revert to prepare for review
