+++ src/e2e/actor.js
@@ -1,16 +1,10 @@
-//import { Actor, HttpAgent } from "@dfinity/agent";
-//import fetch from "isomorphic-fetch";
-//import canisterIds from ".dfx/local/canister_ids.json";
-//import { idlFactory } from "../declarations/PersonalWebSpace_backend/PersonalWebSpace_backend.did.js";
-import {
-  //PersonalWebSpace_backend,
-  createActor as createBackendCanisterActor,
-  canisterId as backendCanisterId,
- // idlFactory as backendIdlFactory,
-} from "../declarations/PersonalWebSpace_backend";
-//import { identity } from "./identity.ts";
+import { Actor, HttpAgent } from "@dfinity/agent";
+import fetch from "isomorphic-fetch";
+import canisterIds from ".dfx/local/canister_ids.json";
+import { idlFactory } from "../declarations/hello/hello.did.js";
+import { identity } from "./identity.ts";
 
-/* export const createActor = async (canisterId, options) => {
+export const createActor = async (canisterId, options) => {
   const agent = new HttpAgent({ ...options?.agentOptions });
   await agent.fetchRootKey();
 
@@ -25,14 +19,6 @@ import {
 export const helloCanister = canisterIds.hello.local;
 
 export const hello = await createActor(helloCanister, {
-  agentOptions: {
-    host: "http://127.0.0.1:8000",
-    fetch,
-    identity: await identity,
-  },
-}); */
-
-export const hello = createBackendCanisterActor(backendCanisterId, {
   agentOptions: {
     host: "http://127.0.0.1:8000",
     fetch,
