+++ src/e2e/actor_changed.js
@@ -1,0 +1,41 @@
+//import { Actor, HttpAgent } from "@dfinity/agent";
+//import fetch from "isomorphic-fetch";
+//import canisterIds from ".dfx/local/canister_ids.json";
+//import { idlFactory } from "../declarations/PersonalWebSpace_backend/PersonalWebSpace_backend.did.js";
+import {
+  //PersonalWebSpace_backend,
+  createActor as createBackendCanisterActor,
+  canisterId as backendCanisterId,
+ // idlFactory as backendIdlFactory,
+} from "../declarations/PersonalWebSpace_backend";
+//import { identity } from "./identity.ts";
+
+/* export const createActor = async (canisterId, options) => {
+  const agent = new HttpAgent({ ...options?.agentOptions });
+  await agent.fetchRootKey();
+
+  // Creates an actor with using the candid interface and the HttpAgent
+  return Actor.createActor(idlFactory, {
+    agent,
+    canisterId,
+    ...options?.actorOptions,
+  });
+};
+
+export const helloCanister = canisterIds.hello.local;
+
+export const hello = await createActor(helloCanister, {
+  agentOptions: {
+    host: "http://127.0.0.1:8000",
+    fetch,
+    identity: await identity,
+  },
+}); */
+
+export const hello = createBackendCanisterActor(backendCanisterId, {
+  agentOptions: {
+    host: "http://127.0.0.1:8000",
+    fetch,
+    identity: await identity,
+  },
+});
