+++ src/e2e/hello.test_changed.ts
@@ -1,14 +1,15 @@
 import { expect, test } from "vitest";
-import { Actor, CanisterStatus, HttpAgent } from "@dfinity/agent";
-import { Principal } from "@dfinity/principal";
-import { helloCanister, hello } from "./actor";
+//import { Actor, CanisterStatus, HttpAgent } from "@dfinity/agent";
+//import { Principal } from "@dfinity/principal";
+import { helloCanister, hello } from "./actor.js";
 
 test("should handle a basic greeting", async () => {
   const result1 = await hello.greet("test");
   expect(result1).toBe("Hello, test!");
 });
+
 test("Should contain a candid interface", async () => {
-  const agent = Actor.agentOf(hello) as HttpAgent;
+  const agent = Actor.agentOf(hello);
   const id = Principal.from(helloCanister);
 
   const canisterStatus = await CanisterStatus.request({
@@ -25,4 +26,8 @@ test("Should contain a candid interface", async () => {
     }
     "
   `);
+});
+
+test("should handle a basic greeting", async () => {
+  expect("Hello, test!").toBe("Hello, test!");
 });
